Prefer uncheckedArgument over argument where possible for callFrame
This saves us the extra work of an if-statement check.

+++ Source/JavaScriptCore/runtime/FunctionPrototype.cpp
@@ -114,8 +114,8 @@ JSC_DEFINE_HOST_FUNCTION(functionProtoFuncBind, (JSGlobalObject* globalObject, C
     JSObject* target = asObject(thisValue);
 
     JSValue boundThis = callFrame->argument(0);
-    unsigned argumentCount = callFrame->argumentCount();
-    unsigned numBoundArgs = 0;
+    size_t argumentCount = callFrame->argumentCount();
+    size_t numBoundArgs = 0;
     ArgList boundArgs { };
     if (argumentCount > 1)
         boundArgs = ArgList(callFrame, 1);
